Bootstrap improvements

+++ bootstrap.sh
@@ -1,4 +1,6 @@
-#!/usr/bin/env bash
+#!/bin/bash -i
+
+set -e
 
 # install base deps
 sudo apt upgrade --assume-yes
@@ -6,7 +8,9 @@ sudo apt install --assume-yes git curl software-properties-common python-is-pyth
 mkdir -p /tmp/dotfiles
 
 # install chezmoi
-wget --no-clobber https://github.com/twpayne/chezmoi/releases/download/v1.8.11/chezmoi_1.8.11_linux_amd64.deb -O /tmp/dotfiles/chezmoi.deb
+if [ ! -f /tmp/dotfiles/chezmoi.deb ] ; then
+    wget --no-clobber https://github.com/twpayne/chezmoi/releases/download/v1.8.11/chezmoi_1.8.11_linux_amd64.deb -O /tmp/dotfiles/chezmoi.deb
+fi
 sudo apt install /tmp/dotfiles/chezmoi.deb
 
 # init and apply chezmoi
@@ -19,9 +23,14 @@ chezmoi apply
 
 # install ansible galaxy dependencies
 ansible-galaxy collection install community.general
-wget --no-clobber -O /tmp/dotfiles/requirements.yml https://raw.githubusercontent.com/alanchrt/dotfiles/regolith/requirements.yml
+if [ ! -f /tmp/dotfiles/requirements.yml ] ; then
+    wget --no-clobber -O /tmp/dotfiles/requirements.yml https://raw.githubusercontent.com/alanchrt/dotfiles/regolith/requirements.yml
+fi
 ansible-galaxy install -r /tmp/dotfiles/requirements.yml
+
+# clean up
+rm -r /tmp/dotfiles
 
 # run ansible playbook
 # test locally: ansible-playbook --ask-become-pass -c local local.yml -i hosts
-ansible-pull --checkout regolith --url git@github.com:alanchrt/dotfiles.git --ask-become-pass
+ANSIBLE_FORCE_COLOR=true ansible-pull --checkout regolith --url git@github.com:alanchrt/dotfiles.git --ask-become-pass -i hosts
